APUSH Game
The game was created to make fun of the experiences my friends and I
went through after taking AP US History. It was intended to make fun of
the struggles everyone went through in the vigorous course. The goal of
the game is to evade all the obstacles and collect all energy drinks
and books on each level. Then on the third level hand everything in to
the teacher